requirements.txt
flask
pandas
google-genai
python-dotenv
flask-limiter